refactor: Improve code formatting and readability in material and translation modules
