+++ nusamai-czml/src/models/czml_polygon.rs
@@ -188,15 +188,21 @@ fn is_default_fill(fill: &CzmlBoolean) -> bool {
 }
 
 fn default_material() -> Material {
-    Material::Object(MaterialProperties {
-        solid_color: Some(SolidColorMaterial {
-            color: Color::Object(ColorProperties {
-                rgba: Some(RgbaValue::Constant([255, 255, 255, 255])),
-                ..Default::default()
-            }.into()),
-        }),
-        ..Default::default()
-    }.into())
+    Material::Object(
+        MaterialProperties {
+            solid_color: Some(SolidColorMaterial {
+                color: Color::Object(
+                    ColorProperties {
+                        rgba: Some(RgbaValue::Constant([255, 255, 255, 255])),
+                        ..Default::default()
+                    }
+                    .into(),
+                ),
+            }),
+            ..Default::default()
+        }
+        .into(),
+    )
 }
 
 fn is_default_material(material: &Material) -> bool {
@@ -212,10 +218,13 @@ fn is_default_outline(outline: &CzmlBoolean) -> bool {
 }
 
 fn default_outline_color() -> Color {
-    Color::Object(ColorProperties {
-        rgba: Some(RgbaValue::Constant([0, 0, 0, 255])),
-        ..Default::default()
-    }.into())
+    Color::Object(
+        ColorProperties {
+            rgba: Some(RgbaValue::Constant([0, 0, 0, 255])),
+            ..Default::default()
+        }
+        .into(),
+    )
 }
 
 fn is_default_outline_color(outline_color: &Color) -> bool {
