+++ nusamai-czml/src/models/model.rs
@@ -95,10 +95,13 @@ fn default_height_reference() -> HeightReference {
 }
 
 fn default_silhouette_color() -> Color {
-    Color::Object(ColorProperties {
-        rgba: Some(RgbaValue::Constant([255, 0, 0, 0])),
-        ..Default::default()
-    }.into())
+    Color::Object(
+        ColorProperties {
+            rgba: Some(RgbaValue::Constant([255, 0, 0, 0])),
+            ..Default::default()
+        }
+        .into(),
+    )
 }
 
 fn default_silhouette_size() -> CzmlDouble {
@@ -106,10 +109,13 @@ fn default_silhouette_size() -> CzmlDouble {
 }
 
 fn default_color() -> Color {
-    Color::Object(ColorProperties {
-        rgba: Some(RgbaValue::Constant([0, 0, 0, 0])),
-        ..Default::default()
-    }.into())
+    Color::Object(
+        ColorProperties {
+            rgba: Some(RgbaValue::Constant([0, 0, 0, 0])),
+            ..Default::default()
+        }
+        .into(),
+    )
 }
 
 fn default_color_blend_mode() -> ColorBlendMode {
