+++ nusamai-czml/src/models/scale.rs
@@ -36,14 +36,17 @@ pub struct ScaleProperties {
 impl Default for Scale {
     fn default() -> Self {
         Self {
-            value: ScaleValueType::Object(ScaleProperties {
-                cartesian: Some(Cartesian3Value::Constant([1.0, 1.0, 1.0])),
-                reference: None,
-                interpolatable_property: None,
-                deletable_property: None,
-                distance_display_condition_value_property: None,
-                reference_value_property: None,
-            }.into()),
+            value: ScaleValueType::Object(
+                ScaleProperties {
+                    cartesian: Some(Cartesian3Value::Constant([1.0, 1.0, 1.0])),
+                    reference: None,
+                    interpolatable_property: None,
+                    deletable_property: None,
+                    distance_display_condition_value_property: None,
+                    reference_value_property: None,
+                }
+                .into(),
+            ),
         }
     }
 }
