+++ nusamai-czml/src/models/stripe_material.rs
@@ -42,17 +42,23 @@ fn default_orientation() -> StripeOrientation {
 }
 
 fn default_even_color() -> Color {
-    Color::Object(ColorProperties {
-        rgba: Some(RgbaValue::Constant([255, 255, 255, 0])),
-        ..Default::default()
-    }.into())
+    Color::Object(
+        ColorProperties {
+            rgba: Some(RgbaValue::Constant([255, 255, 255, 0])),
+            ..Default::default()
+        }
+        .into(),
+    )
 }
 
 fn default_odd_color() -> Color {
-    Color::Object(ColorProperties {
-        rgba: Some(RgbaValue::Constant([0, 0, 0, 0])),
-        ..Default::default()
-    }.into())
+    Color::Object(
+        ColorProperties {
+            rgba: Some(RgbaValue::Constant([0, 0, 0, 0])),
+            ..Default::default()
+        }
+        .into(),
+    )
 }
 
 fn default_offset() -> CzmlDouble {
