+++ nusamai-czml/src/models/translation.rs
@@ -36,14 +36,17 @@ pub struct TranslationProperties {
 impl Default for Translation {
     fn default() -> Self {
         Self {
-            value: TranslationValueType::Object(TranslationProperties {
-                cartesian: Some(Cartesian3Value::Constant([0.0, 0.0, 0.0])),
-                reference: None,
-                interpolatable_property: None,
-                deletable_property: None,
-                distance_display_condition_value_property: None,
-                reference_value_property: None,
-            }.into()),
+            value: TranslationValueType::Object(
+                TranslationProperties {
+                    cartesian: Some(Cartesian3Value::Constant([0.0, 0.0, 0.0])),
+                    reference: None,
+                    interpolatable_property: None,
+                    deletable_property: None,
+                    distance_display_condition_value_property: None,
+                    reference_value_property: None,
+                }
+                .into(),
+            ),
         }
     }
 }
